+++ lib/scss_lint/linter/unnecessary_mantissa.rb
@@ -9,6 +9,7 @@ module SCSSLint
     def visit_script_string(node)
       return unless node.type == :identifier
       return if node.value =~ /^'|"/
+      return if url_literal?(node)
 
       node.value.scan(REAL_NUMBER_REGEX) do |number, integer, mantissa, units|
         if unnecessary_mantissa?(mantissa)
@@ -40,6 +41,10 @@ module SCSSLint
 
     def unnecessary_mantissa?(mantissa)
       mantissa !~ /[^0]/
+    end
+
+    def url_literal?(node)
+      node.value.start_with?('url(')
     end
   end
 end